+++ build.gradle
@@ -9,7 +9,7 @@ plugins {
 }
 
 application {
-    mainClass.set(System.getProperty('main_class') ?: 'Main')
+    mainClass.set(System.getProperty('main_class') ?: 'MyCompiler')
 }
 
 repositories {
@@ -17,9 +17,11 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.slf4j:slf4j-simple:2.0.17'
     implementation 'org.apache.commons:commons-text:1.15.0'
     implementation 'org.knowm.xchart:xchart:3.8.8'
+    implementation 'org.slf4j:slf4j-simple:2.0.17'
+    implementation 'io.javalin:javalin:7.2.0'
+    implementation 'org.json:json:20251224'
 }
 
 java {
