+++ src/main/java/MySupplier.java
@@ -1,31 +1,11 @@
 public interface MySupplier {
-  /**
-   * Returns the number of series to be generated.
-   *
-   * @return the number of series
-   */
-  int numberOfSeries();
+  int numberOfCharts();
 
-  /**
-   * Returns the title for a specific series.
-   *
-   * @param seriesIndex the index of the series
-   * @return the title of the series
-   */
-  String getTitle(int seriesIndex);
+  int numberOfSeries(int chartNumber);
 
-  /**
-   * Generates a series of data points for a specific series.
-   *
-   * @param seriesIndex the index of the series
-   * @return a 2D array where each row contains [x, y] values for the series
-   */
-  double[][] generateSeries(int seriesIndex);
+  String getTitle(int chartNumber, int seriesIndex);
 
-  /**
-   * Generates a plot of the series data.
-   *
-   * @return a base64 encoded string representing the plot image
-   */
+  double[][] generateSeries(int chartNumber, int seriesIndex);
+
   String plot();
 }
