+++ src/main/java/MySupplierImplOrig.java
@@ -1,0 +1,86 @@
+import org.json.JSONArray;
+import org.knowm.xchart.BitmapEncoder;
+import org.knowm.xchart.XYChart;
+import org.knowm.xchart.XYChartBuilder;
+
+public class MySupplierImplOrig implements MySupplier {
+  private static final int[][] data = {
+    {1, 2},
+    {3, 4},
+    {5, 6},
+    {3, 4},
+    {4, 3}
+  };
+
+  public int numberOfCharts() {
+    return 2;
+  }
+
+  public int numberOfSeries(int chartNumber) {
+    return (new int[] {1, 2})[chartNumber];
+  }
+
+  public String getTitle(int chartNumber, int seriesIndex) {
+    return "S_" + (chartNumber + 1) + "_" + (seriesIndex + 1);
+  }
+
+  public double[][] generateSeries(int chartNumber, int seriesIndex) {
+    double[][] series = new double[data.length][2];
+    if (chartNumber == 0) {
+      double base = data[0][0];
+      for (int i = 0; i < series.length; i++) {
+        double x = i + 1;
+        double y = (data[i][0] / base) * 100.0;
+        series[i][0] = x;
+        series[i][1] = y;
+      }
+      return series;
+    }
+    if (chartNumber == 1) {
+      double baseA = data[0][0];
+      double baseB = data[0][1];
+      for (int i = 0; i < series.length; i++) {
+        double x = i + 1;
+        double y =
+            seriesIndex == 0 ? ((data[i][0] / baseA) * 100.0) : ((data[i][1] / baseB) * 100.0);
+        series[i][0] = x;
+        series[i][1] = y;
+      }
+      return series;
+    }
+    return null;
+  }
+
+  // Example function
+  private double f(double x, double a) {
+    return x * x * a;
+  }
+
+  // Do not change this method, unless you know what you are doing.
+  // It is used to plot the series generated by the methods above.
+  public String plot() {
+    JSONArray jsonArray = new JSONArray();
+    for (int c = 0; c < numberOfCharts(); c++) {
+      XYChart chart =
+          new XYChartBuilder()
+              .width(800)
+              .height(400)
+              .title("Chart_" + c)
+              .xAxisTitle("x")
+              .yAxisTitle("y")
+              .build();
+      for (int i = 0; i < numberOfSeries(c); i++) {
+        double[][] series = generateSeries(c, i);
+        double[] xData = new double[series.length];
+        double[] yData = new double[series.length];
+        for (int j = 0; j < series.length; j++) {
+          xData[j] = series[j][0];
+          yData[j] = series[j][1];
+        }
+        chart.addSeries(getTitle(c, i), xData, yData);
+      }
+      jsonArray.put(MyPlotter.imgToBase64String(BitmapEncoder.getBufferedImage(chart)));
+    }
+    return jsonArray.toString();
+  }
+}
