+++ src/main/resources/templates/index.html
@@ -7,43 +7,70 @@
 </head>
 
 <body>
-    <div id="code-editor" style="width: 800px;"></div>
-    <br><br>
-    <input type="button" value="Send Code" onclick="sendCode();">
-    <br><br>
-    <img src="{{1}}" alt="Java Charts Webui" id="chart-img">
-    <script type="module">
-        import { EditorView, basicSetup } from "https://esm.sh/codemirror"
-        import { java } from "https://esm.sh/@codemirror/lang-java"
-        import * as lang from "https://esm.sh/@codemirror/language"
-        let view = new EditorView({
-            parent: document.getElementById("code-editor"),
-            doc: `{{0}}`,
-            extensions: [basicSetup, java()]
+<div id="code-editor" style="width: 800px;"></div>
+<br><br>
+<input type="button" value="Send Code" onclick="sendCode();">
+<div id="result"></div>
+<script type="module">
+    import {EditorView, basicSetup} from "https://esm.sh/codemirror"
+    import {java} from "https://esm.sh/@codemirror/lang-java"
+    import * as lang from "https://esm.sh/@codemirror/language"
+
+    let view = new EditorView({
+        parent: document.getElementById("code-editor"),
+        doc: `{{0}}`,
+        extensions: [basicSetup, java()]
+    });
+
+    foldPlotMethod();
+
+    function sendCode() {
+        fetch("/code", {
+            method: "POST",
+            headers: {'Content-Type': 'application/json'},
+            body: JSON.stringify({code: view.state.doc.toString()})
+        }).then(res => {
+            return res.json();
+        }).then(json => {
+            let r = document.getElementById("result");
+            r.innerHTML = "";
+            for (let i = 0; i < json.length; i++) {
+                let imgSrc = json[i];
+                r.append(document.createElement("br"));
+                r.append(document.createElement("br"));
+                let img = document.createElement("img");
+                img.src = imgSrc;
+                img.id = "chart-img-" + i;
+                r.append(img);
+            }
+            foldPlotMethod();
+            scrollToChart();
         });
-        {
-            // Fold the method plot() in the code:
-            let d = view.state.doc;
-            for (let i = d.lines; i > 0; i--) {
-                let line = d.line(i);
-                if (line.text.includes("public String plot()")) {
-                    let range = lang.foldable(view.state, line.from, line.to);
-                    if (range) {
-                        console.log("Folding ...", range);
-                        view.dispatch({ effects: [lang.foldEffect.of(range)] });
-                    }
-                    break;
+    }
+
+    window.sendCode = sendCode;
+
+    function foldPlotMethod() {
+        // Fold the method plot() in the code:
+        let d = view.state.doc;
+        for (let i = d.lines; i > 0; i--) {
+            let line = d.line(i);
+            if (line.text.includes("public String plot()")) {
+                let range = lang.foldable(view.state, line.from, line.to);
+                if (range) {
+                    console.log("Folding ...", range);
+                    view.dispatch({effects: [lang.foldEffect.of(range)]});
                 }
+                break;
             }
         }
+    }
+
+    function scrollToChart() {
         // Scroll to the chart image after the editor is ready
-        document.getElementById("chart-img").scrollIntoView({ block: "start", behavior: "smooth" });
-        function sendCode() {
-            let code = encodeURIComponent(view.state.doc.toString());
-            window.location.href = "/code/" + code;
-        }
-        window.sendCode = sendCode;
-    </script>
+        document.getElementById("chart-img-0").scrollIntoView({behavior: "smooth"});
+    }
+</script>
 </body>
 
 </html>
